"+ Event"
Added “+ Event” button. This was created last week, but had did not
commit because I was trying to also add the “+ User”. For reason’s
unknown the the webpage was only letting one of the buttons show up.
Also, when adding the form functionality to the button the entire
calendar vanished..